# openapi.yaml
openapi: 3.0.0
info:
  title: Configurações API
  description: API para otimização de configurações de componentes de computadores
  version: 1.0.0
paths:
  /api/melhor_configuracao:
    post:
      summary: Gera e salva a melhor configuração
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                limite_valor:
                  type: number
      responses:
        "200":
          description: Configuração salva com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  melhor_configuracao:
                    type: object
                  custo_final:
                    type: number
                  pontuacao_final:
                    type: number
        "400":
          description: Erro nos dados fornecidos

  /api/configuracoes:
    get:
      summary: Recupera todas as configurações
      responses:
        "200":
          description: Lista de configurações
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    configuracao:
                      type: object
                    custo_final:
                      type: number
                    pontuacao_final:
                      type: number
                    created_at:
                      type: string
                      format: date-time

  /api/configuracoes/{id}:
    get:
      summary: Recupera uma configuração específica pelo ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Detalhes da configuração
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  configuracao:
                    type: object
                  custo_final:
                    type: number
                  pontuacao_final:
                    type: number
                  created_at:
                    type: string
                    format: date-time
        "404":
          description: Configuração não encontrada
    delete:
      summary: Remove uma configuração pelo ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Configuração removida com sucesso
        "404":
          description: Configuração não encontrada
    put:
      summary: Atualiza uma configuração pelo ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                configuracao:
                  type: object
                custo_final:
                  type: number
                pontuacao_final:
                  type: number
      responses:
        "200":
          description: Configuração alterada com sucesso
        "400":
          description: Erro nos dados fornecidos
        "404":
          description: Configuração não encontrada